requirements.txt
flask
requests
pytest
minio
python-dotenv